added tests for torch embedders
